% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix_function.R
\name{cor_matrix}
\alias{cor_matrix}
\title{cor_matrix_function}
\usage{
cor_matrix(data, year_val)
}
\arguments{
\item{year_val}{Which year do you want to set.}

\item{dataset}{}
}
\description{
This function allows you to get the correlation table among the numerical variables within the selected year
}
\examples{
cor_matrix()
}
\keyword{correlation,}
\keyword{matrix}
