% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player_function.R
\name{oldest_player}
\alias{oldest_player}
\title{oldest_player_function}
\usage{
oldest_player(data, year_val)
}
\arguments{
\item{data}{The dataset you are loading in this function}

\item{year_val}{Which year do you want to set.}
}
\description{
This function allows you get the total points made at specific year when your input the dataset and the year you want.
}
\examples{
oldest_player()
oldest_player(data,year_val)
}
\keyword{nba,}
\keyword{oldest}
\keyword{player}
