% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youngest_player_function.R
\name{youngest_player}
\alias{youngest_player}
\title{youngest_player_function}
\usage{
youngest_player(data, year_val)
}
\arguments{
\item{year_val}{Which year do you want to set.}
}
\description{
This function allows you get the youngest player  at specific year when your input the dataset and the year you want.
}
\examples{
youngest_player()

}
\keyword{nba,}
\keyword{player}
\keyword{youngest}
